requirements.txt
albumentations==1.3.0
einops==0.3.0
fvcore==0.1.5.post20221221
gradio==3.39.0
numpy==1.23.1
omegaconf==2.1.1
open_clip_torch==2.17.1
opencv_contrib_python==4.3.0.36
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.4.0
pytorch_lightning==1.5.0
safetensors==0.2.7
scipy==1.9.1
setuptools==66.0.0
share==1.0.4
submitit==1.5.1
timm==0.6.12
torch==2.0.0
torchmetrics==0.6.0
tqdm==4.65.0
transformers==4.19.2
xformers==0.0.18
pyemd
geomloss
kornia